requirements.txt
click==7.1.1
dominate==2.5.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.3
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
pkg-resources==0.0.0
PyYAML==5.3.1
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.2.1
